LPDEV-597: finish up with clean design in filters, invoices

+++ IATA/Modules/Network/Models/Invoice/Request/PaymentRequest.swift
@@ -8,7 +8,7 @@ class PaymentRequest:  BaseMappable {
     }
     
     var invoicesIds: [String]? = []
-    var amountInBaseAsset: Int?
+    var amountInBaseAsset: Double?
     
     internal required init?() {
     }
